# gh-74690: Improve performance when testing against protocols with many attributes.